S12P11A204/hw/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-motor==3.4.14
adafruit-circuitpython-pca9685==3.4.16
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-servokit==1.3.18
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.76.1
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.5.2
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.13.3
binho-host-adapter==0.1.6
bleach==6.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
Click==7.0
colorama==0.4.3
comm==0.2.2
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.8.12
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httpx==0.28.1
idna==2.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jetson-MFRC522==1.0.0
jetson-stats==4.3.1
Jetson.GPIO==2.1.6
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
-e git+https://github.com/jaybdub/jupyter_clickable_image_widget@b2ee0f45e6ec317f2f68f7905388877d5ec6c486#egg=jupyter_clickable_image_widget
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.1
monotonic==1.5
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
overrides==7.7.0
packaging==24.2
PAM==0.4.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.6.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.16.2
pycparser==2.22
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.55.4
Pygments==2.19.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynmea2==1.19.0
pynput==1.7.7
pyparsing==2.4.6
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-dbusmock==0.19
python-debian==0.1.36+ubuntu1.1
python-json-logger==3.2.1
python-xlib==0.33
pytz==2024.2
pyusb==1.2.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.2.1
readchar==4.2.1
referencing==0.35.1
requests==2.32.3
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.3
simplejson==3.16.0
six==1.14.0
smbus2==0.5.0
sniffio==1.3.1
soupsieve==2.6
sparkfun-pi-servo-hat==0.9.1
sparkfun-qwiic-adxl313==0.0.7
sparkfun-qwiic-alphanumeric==0.0.1
sparkfun-qwiic-as6212==0.0.2
sparkfun-qwiic-bme280==0.9.0
sparkfun-qwiic-button==2.0.1
sparkfun-qwiic-ccs811==0.9.4
sparkfun-qwiic-dual-encoder-reader==0.0.2
sparkfun-qwiic-eeprom==0.0.1
sparkfun-qwiic-gpio==0.0.2
sparkfun-qwiic-icm20948==0.0.1
sparkfun-qwiic-joystick==0.9.0
sparkfun-qwiic-keypad==0.9.0
sparkfun-qwiic-kx13x==1.0.1
sparkfun-qwiic-large-oled==0.0.1
sparkfun-qwiic-led-stick==0.0.1
sparkfun-qwiic-max3010x==0.0.2
sparkfun-qwiic-micro-oled==0.10.0
sparkfun-qwiic-oled-base==0.0.2
sparkfun-qwiic-oled-display==0.0.2
sparkfun-qwiic-pca9685==0.9.1
sparkfun-qwiic-pir==0.0.4
sparkfun-qwiic-proximity==0.9.0
sparkfun-qwiic-relay==0.0.2
sparkfun-qwiic-rfid==2.0.0
sparkfun-qwiic-scmd==0.9.1
sparkfun-qwiic-serlcd==0.0.1
sparkfun-qwiic-sgp40==0.0.4
sparkfun-qwiic-soil-moisture-sensor==0.0.2
sparkfun-qwiic-tca9548a==0.9.0
sparkfun-qwiic-titan-gps==0.1.1
sparkfun-qwiic-twist==0.9.0
sparkfun-qwiic-vl53l1x==1.0.1
sparkfun-top-phat-button==0.0.2
sparkfun-ublox-gps==1.1.5
sparkfun_qwiic==1.1.10
sparkfun_qwiic_i2c==1.0.0
sparkfun_qwiic_otos==0.0.1
spidev==3.6
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
sysv-ipc==1.1.0
terminado==0.18.1
tinycss2==1.2.1
tomli==2.2.1
torch==2.1.0a0+41361538.nv23.6
torchaudio==2.2.0
torchvision==0.16.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transitions==0.9.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ultralytics==8.3.61
ultralytics-thop==2.0.13
uri-template==1.3.0
urllib3==1.25.8
urwid==2.0.1
wadllib==1.3.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
xkit==0.0.0
yarl==1.15.2
zipp==3.20.2
